Convert TAP fonts and UDG to binary numbers source
The goal is to arrange the the binary grids of the graphics into blocks
and sprites, in order to modify them easily. Then Makefile will convert
them to `c,` Forth format to integrate them into the main source.
FantomoUDG (http://programandala.net/en.program.fantomoudg.html) was
used to do the conversion.