Improve movement and debugging code
The code that manages the movement on the sea and on the island has been
factored, tidied and renamed. Not fully tested though.

The debug tools that print the sea and island maps has been improved to
show the grid in right order and also the current position of the
player.